+++ backend/schemas/users.py
@@ -1,5 +1,15 @@
-from pydantic import BaseModel
+from pydantic import BaseModel, EmailStr
 
-class Token(BaseModel):
-    access_token: str
-    token_type: str
+
+class UserCreate(BaseModel):
+    username: str
+    email: EmailStr
+    password: str
+
+
+class ShowUser(BaseModel):
+    username: str
+    email: EmailStr
+
+    class Config:
+        orm_mode = True
