+++ requirements.txt
@@ -11,7 +11,9 @@ click==8.0.4
 colorama==0.4.4
 cryptography==36.0.2
 distlib==0.3.4
+dnspython==2.2.1
 ecdsa==0.17.0
+email-validator==1.1.3
 fastapi==0.75.0
 filelock==3.6.0
 greenlet==1.1.2
@@ -30,6 +32,7 @@ py==1.11.0
 pyasn1==0.4.8
 pycparser==2.21
 pydantic==1.9.0
+pydantic[email]
 PyMySQL==1.0.2
 pyparsing==3.0.7
 pytest==7.1.1
